move config class to a separate file
