+++ snuba/clickhouse/config.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+from typing import Optional
+
+
+@dataclass(frozen=True)
+class ClickhouseConnectionConfig:
+    host: str
+    port: int
+    http_port: Optional[int]
