+++ snuba/datasets/cdc/groupedmessage.py
@@ -1,6 +1,7 @@
 from typing import Sequence
 
 from snuba.clickhouse.columns import ColumnSet, DateTime, Nullable, UInt
+from snuba.clickhouse.config import ClickhouseConnectionConfig
 from snuba.clickhouse.pool import ClickhousePool
 
 from snuba.datasets.cdc import CdcDataset
@@ -15,7 +16,6 @@ from snuba.datasets.table_storage import TableWriter, KafkaStreamLoader
 from snuba.query.processors.basic_functions import BasicFunctionsProcessor
 from snuba.query.processors.prewhere import PrewhereProcessor
 from snuba.query.query_processor import QueryProcessor
-from snuba.settings import ClickhouseConnectionConfig
 from snuba.snapshots.loaders.single_table import SingleTableBulkLoader
